+++ src/methods/property_solvers/multicomponent/flash/PT.jl
@@ -21,12 +21,12 @@ Outputs - Tuple containing:
  - nᵢⱼ, Array of mole numbers of species j in phase i, [mol]
  - G, Gibbs Free Energy of Equilibrium Mixture [J]
 """
-function tp_flash(model::EoSModel, p, T, n; kwargs...)
+function tp_flash(model::EoSModel, p, T, n = SA[1.0]; kwargs...)
     method = init_preferred_method(tp_flash,model,kwargs)
     return tp_flash(model, p, T, n, method)
 end
 
-function tp_flash(model::EoSModel, p, T, n,method::FlashMethod)
+function tp_flash(model::EoSModel, p, T, n, method::FlashMethod)
     result = tp_flash2(model, p, T, n,method)
     return tp_flash2_to_tpflash(model,p,T,n,result)
 end