+++ src/methods/property_solvers/multicomponent/fugacity.jl
@@ -1,8 +1,21 @@
 ##general successive substitution method to solve bubble/dew problems via fugacity coefficients
+module FugEnum
+    @enum BubbleDew begin
+        BUBBLE_PRESSURE
+        BUBBLE_TEMPERATURE
+        DEW_PRESSURE
+        DEW_TEMPERATURE
+    end
+    is_bubble(x::BubbleDew) = (x == BUBBLE_PRESSURE || x == BUBBLE_TEMPERATURE)
+    is_dew(x::BubbleDew) = !is_bubble(x)
+    is_temperature(x::BubbleDew) = (x == BUBBLE_TEMPERATURE  || x == DEW_TEMPERATURE)
+    is_pressure(x::BubbleDew) = !is_temperature(x) 
+end
 
-function _fug_OF_ss(model::EoSModel,p,T,x,y,vol0,_bubble,_pressure;itmax_ss = 5, itmax_newton = 10,tol_pT = 1e-8,tol_xy = 1e-8,tol_of = 1e-8)
+function _fug_OF_ss(model::EoSModel,p,T,x,y,vol0,method;itmax_ss = 5, itmax_newton = 10,tol_pT = 1e-8,tol_xy = 1e-8,tol_of = 1e-8)
     volx,voly = vol0
     converged = false
+    _bubble,_pressure = FugEnum.is_bubble(method),FugEnum.is_pressure(method)
     tol_stability = abs2(cbrt(tol_xy))
     #caches for ∂lnϕ∂n∂P∂T/∂lnϕ∂n∂P
     if _pressure
@@ -141,21 +154,22 @@ function _fug_OF_ss(model::EoSModel,p,T,x,y,vol0,_bubble,_pressure;itmax_ss = 5,
     return converged,(p,T,_x,_y,(volx,voly),lnK)
 end
 
-function _fug_OF_ss(modelx::EoSModel,modely::Nothing,p,T,x,y,vol0,_bubble,_pressure,_view;itmax_ss = 5, itmax_newton = 10,tol_pT = 1e-8,tol_xy = 1e-8,tol_of = 1e-8)
-    return _fug_OF_ss(modelx,p,T,x,y,vol0,_bubble,_pressure;itmax_ss, itmax_newton,tol_pT,tol_xy,tol_of)
+function _fug_OF_ss(modelx::EoSModel,modely::Nothing,p,T,x,y,vol0,method,_view;itmax_ss = 5, itmax_newton = 10,tol_pT = 1e-8,tol_xy = 1e-8,tol_of = 1e-8)
+    return _fug_OF_ss(modelx,p,T,x,y,vol0,method;itmax_ss, itmax_newton,tol_pT,tol_xy,tol_of)
 end
 
-function _fug_OF_ss(modelx::Nothing,modely::EoSModel,p,T,x,y,vol0,_bubble,_pressure,_view;itmax_ss = 5, itmax_newton = 10,tol_pT = 1e-8,tol_xy = 1e-8,tol_of = 1e-8)
-    return _fug_OF_ss(modely,p,T,x,y,vol0,_bubble,_pressure;itmax_ss, itmax_newton,tol_pT,tol_xy,tol_of)
+function _fug_OF_ss(modelx::Nothing,modely::EoSModel,p,T,x,y,vol0,method,_view;itmax_ss = 5, itmax_newton = 10,tol_pT = 1e-8,tol_xy = 1e-8,tol_of = 1e-8)
+    return _fug_OF_ss(modely,p,T,x,y,vol0,method;itmax_ss, itmax_newton,tol_pT,tol_xy,tol_of)
 end
 
 
 ##general successive substitution method to solve bubble/dew problems via fugacity coefficients
 ##support for nonvolatiles/noncondensables
 
-function _fug_OF_ss(modelx::EoSModel,modely::EoSModel,p,T,x,y,vol0,_bubble,_pressure,_view;itmax_ss = 5, itmax_newton = 10,tol_pT = 1e-8,tol_xy = 1e-8,tol_of = 1e-8)
+function _fug_OF_ss(modelx::EoSModel,modely::EoSModel,p,T,x,y,vol0,method,_view;itmax_ss = 5, itmax_newton = 10,tol_pT = 1e-8,tol_xy = 1e-8,tol_of = 1e-8)
     volx,voly = vol0
     converged = false
+    _bubble,_pressure = FugEnum.is_bubble(method),FugEnum.is_pressure(method)
     tol_stability = abs2(cbrt(tol_xy))
     if _bubble
         n = length(modely)
@@ -375,7 +389,10 @@ function _fug_J(J,x,y,∂lnϕ∂nx,∂lnϕ∂ny,_bubble)
     return J
 end
 
-function _fug_OF_neqsystem(model,_x, _y, _p, _T, vol_cache,_bubble,_pressure,_phase)
+function _fug_OF_neqsystem(model,_x, _y, _p, _T, vol_cache,method,_phase)
+    
+    _bubble,_pressure = FugEnum.is_bubble(method),FugEnum.is_pressure(method)
+ 
     #caches for ∂lnϕ∂n∂P∂T/∂lnϕ∂n∂P
 
     XX = something(_p,_T)
@@ -458,12 +475,12 @@ end
 ##support for noncondensables/nonvolatiles
 
 #dispatch to simplified function
-function _fug_OF_neqsystem(modelx::EoSModel,modely::Nothing,_x, _y, _p, _T, vol_cache,_bubble,_pressure,_phase,_view)
-    return  _fug_OF_neqsystem(modelx,_x, _y, _p, _T, vol_cache,_bubble,_pressure,_phase)
+function _fug_OF_neqsystem(modelx::EoSModel,modely::Nothing,_x, _y, _p, _T, vol_cache,method,_phase,_view)
+    return  _fug_OF_neqsystem(modelx,_x, _y, _p, _T, vol_cache,method,_phase)
 end
 
-function _fug_OF_neqsystem(modelx::Nothing,modely::EoSModel,_x, _y, _p, _T, vol_cache,_bubble,_pressure,_phase,_view)
-    return  _fug_OF_neqsystem(modely,_x, _y, _p, _T, vol_cache,_bubble,_pressure,_phase)
+function _fug_OF_neqsystem(modelx::Nothing,modely::EoSModel,_x, _y, _p, _T, vol_cache,method,_phase,_view)
+    return  _fug_OF_neqsystem(modely,_x, _y, _p, _T, vol_cache,method,_phase)
 end
 
 
@@ -481,25 +498,26 @@ function _select_xy(w,K,x,y,_bubble,_view)
 end
 
 #main function
-function _fug_OF_neqsystem(modelx::EoSModel,modely::EoSModel,_x, _y, _p, _T, vol_cache,_bubble,_pressure,_phase,_view)
+function _fug_OF_neqsystem(modelx::EoSModel,modely::EoSModel, _x, _y, _p, _T, vol_cache,method,_phase,_view)
+    _bubble,_pressure = FugEnum.is_bubble(method),FugEnum.is_pressure(method)
     if _bubble
-        w = similar(_y)
-        wx = x
+        w = similar(@view(_x[_view]))
+        wx = _x
         wy = w
     else
-        w = similar(_x)
+        w = similar(@view(_y[_view]))
         wx = w
-        wy = y
+        wy = _y
     end
 
     K = similar(w)
     XX = something(_p,_T)
     if _pressure
-        Hϕx = ∂lnϕ_cache(model, XX, XX, wx, Val{false}())
-        Hϕy = ∂lnϕ_cache(model, XX, XX, wy, Val{false}())
+        Hϕx = ∂lnϕ_cache(modelx, XX, XX, wx, Val{false}())
+        Hϕy = ∂lnϕ_cache(modely, XX, XX, wy, Val{false}())
     else
-        Hϕx = ∂lnϕ_cache(model, XX, XX, wx, Val{true}())
-        Hϕy = ∂lnϕ_cache(model, XX, XX, wy, Val{true}())
+        Hϕx = ∂lnϕ_cache(modelx, XX, XX, wx, Val{true}())
+        Hϕy = ∂lnϕ_cache(modely, XX, XX, wy, Val{true}())
     end
 
     function f!(F, inc)