+++ test/test_methods_api_flash.jl
@@ -93,7 +93,7 @@
         #water-oxygen system, non-condensables
         model_a_ideal = CompositeModel(["water","oxygen"],liquid = RackettLiquid,gas = BasicIdeal,saturation = DIPPR101Sat)
         @test Clapeyron.tp_flash(model_a_ideal,134094.74892634258,70 + 273.15,[18500.0, 24.08],noncondensables = ["oxygen"])[1] ≈
-        [1.0 0.0; 
+        [1.0 0.0;
         0.23252954843762222 0.7674704515623778] rtol = 1e-6
     end
 
@@ -279,7 +279,7 @@ end
     fluids= ["isopentane","isobutane"]
     model = cPR(fluids,idealmodel=ReidIdeal)
 
-    p = 2*101325.0; z = [2.0,5.0]; 
+    p = 2*101325.0; z = [2.0,5.0];
     q = 0.062744140625
     res_qp3 = qp_flash(model,q,p,z)
     res_qp4 = qp_flash(model,q,p,z./10)
@@ -294,6 +294,14 @@ end
     n_O2_a = 24.08 # mol O2
     sol_fl = vt_flash(model_a_pr, V_a, T, [n_H2O_a, n_O2_a])
     @test V_a ≈ volume(sol_fl)
+
+    #PH flash with supercritical pure components (#361)
+    fluid_model = SingleFluid("Hydrogen")
+    T_in = 70               # K
+    p_in = 350e5           # Pa
+    h_in = enthalpy(fluid_model,p_in,T_in)
+    sol_sc = ph_flash(fluid_model,p_in,h_in)
+    @test Clapeyron.temperature(sol_sc) ≈ T_in
 end
 
 @testset "Saturation Methods" begin
@@ -517,6 +525,9 @@ end
         (Tb,vlb,vvb,yb) = bubble_temperature(system2,p,x0,ChemPotBubbleTemperature(y0 = y0,T0 = T,nonvolatiles = ["decane"]))
         @test Tb  ≈ Tres1 rtol = 1E-6
         @test yb[4] == 0.0
+        #test if the nonvolatile neq system is being built
+        (Tc,vlc,vvc,yc) = bubble_temperature(system2,p,x0,FugBubbleTemperature(itmax_newton = 1, y0 = y0,T0 = T,nonvolatiles = ["decane"]))
+        @test Tc isa Number
     end
     GC.gc()